+++ tests/conftest.py
@@ -114,7 +114,8 @@ def slc_project_class(some_bounds_slc, s1_slc_master, s1_slc_ost_master):
 
         yield s1_batch
     finally:
-        shutil.rmtree(TEMP_SLC_DIR)
+        print('hh')
+        # shutil.rmtree(TEMP_SLC_DIR)
 
 
 @pytest.fixture(scope='session')