+++ tests/test_project.py
@@ -7,8 +7,8 @@ def test_update_ard_param(grd_project_class):
     assert grd_project_class.config_dict["processing"]["single_ARD"]["type"] == 'OST-GTC'
 
 
-# Test GRDs to ARD kind of batch
-def test_grds_to_ard(grd_project_class):
+# # Test GRDs to ARD kind of batch
+def test_grds_to_ards(grd_project_class):
     for ard_type in CONFIG_CHECK['type']['choices']:
         if ard_type in CONFIG_CHECK['ard_types_grd']['choices']:
             grd_project_class.ard_parameters["single_ARD"]["type"] = ard_type
@@ -23,4 +23,22 @@ def test_grds_to_ard(grd_project_class):
                 overwrite=True,
                 to_tif=True,
                 cut_to_aoi=False
+            )
+
+
+# Test Bursts to ARD kind of batch
+def test_bursts_to_ards(slc_project_class):
+    for ard_type in CONFIG_CHECK['type']['choices']:
+        if ard_type in CONFIG_CHECK['ard_types_slc']['choices'] and 'OST-GTC' in ard_type:
+            slc_project_class.ard_parameters["single_ARD"]["type"] = ard_type
+            slc_project_class.update_ard_parameters()
+            slc_project_class.ard_parameters['single_ARD']['resolution'] = 20
+            slc_project_class.update_ard_parameters()
+            # Test ard, Timeseries and Timescan with just one product
+            slc_project_class.bursts_to_ards(
+                timeseries=True,
+                timescan=True,
+                mosaic=False,
+                overwrite=True,
+                max_workers=2
             )